move components in user report profile to incidents
